api_yamdb/requirements.txt
asgiref==3.5.2
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv==0.20.0
requests==2.26.0
requests-oauthlib==1.3.1
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0

